GOT GAME DRAW LOGIC TO FINALLY WORK
Draw function now works properly and when a draw occurs the
status bar is removed and the 'draw-bar' is shown.  Also when a
new game is started draw bar is hidden again and results bar
is shown.

Next need to get API to work, need to fetch number of games played
and store moves in API